move classes to a separate file
